Clean up code, comments, reorganize .jsx files
