+++ frontend/src/App.jsx
@@ -1,10 +1,9 @@
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
-
-import Login from "./pages/Login";
-import Register from "./pages/Register";
-import NotFound from "./pages/NotFound";
+import Login from "./routes/Login";
+import Register from "./routes/Register";
+import NotFound from "./routes/NotFound";
 import ProtectedRoute from "./components/ProtectedRoute";
-import Dashboard from "./pages/Dashboard";
+import Dashboard from "./routes/Dashboard";
 
 function Logout(){
   localStorage.clear();
