+++ frontend/src/pages/AI.jsx
